requirements.txt
--index-url https://pypi.python.org/simple/
Appium-Python-Client==0.17